OPC UA client/server hardening and node values persistency (#245)
* Fix website formatting

* Add support for writing UA data variables, support for reading complex types and removal of more dead code.

* More code cleanup, UI cleanup and implement write data node from the UI.

* Node values hardening and add owner to UI display.

* Add save updated nodeset values functionality.

* Improve OPC UA logging.

* Add session cache to reduce the number of UA servers we're starting and help with persistency.

* Checking for correct user ID during DB operations. Also protecting admin entries as well as entries for everyone.

* Re-arrange landing page to make the most often used items appear first.

* Bug fixes.

* whitespace fix

* Update features description text.

* Fix typo.

* Implemented copy of nodeset and bug fixes. A few UI improvements, too.

* Add correct Razor components.

* Bug fix.

* Code cleanup.

* Fix unit test to properly test paging and also fix CloudLibSync.

* Fix memory leak.

* Update UACloudLibraryServer/UAClientServer/UAClient.cs


* Bug fixes.

* Bug fix.

* Update NuGets only.

* Add retry mechanism for UA server init when Cloud Lib is not running in container and a port is already in use on the host machine.

* whitespace fix.

* Bug fix.

---------

+++ UACloudLibraryServer/Controllers/InfoModelController.cs
@@ -67,7 +67,7 @@ namespace Opc.Ua.Cloud.Library.Controllers
             [FromQuery][SwaggerParameter("Pagination limit")] int? limit
             )
         {
-            UANameSpace[] results = _database.FindNodesets(keywords, offset, limit);
+            UANameSpace[] results = _database.FindNodesets(keywords, User.Identity.Name, offset, limit);
             return new ObjectResult(results) { StatusCode = (int)HttpStatusCode.OK };
         }
 
@@ -76,7 +76,7 @@ namespace Opc.Ua.Cloud.Library.Controllers
         [SwaggerResponse(statusCode: 200, type: typeof(string[]), description: "All OPC UA Information Model namespace URIs and associated identifiers of the models found in the UA Cloud Library.")]
         public async Task<IActionResult> GetAllNamespacesandIdentifiersAsync()
         {
-            string[] results = await _database.GetAllNamespacesAndNodesets().ConfigureAwait(false);
+            string[] results = await _database.GetAllNamespacesAndNodesets(User.Identity.Name).ConfigureAwait(false);
             return new ObjectResult(results) { StatusCode = (int)HttpStatusCode.OK };
         }
 
@@ -85,7 +85,7 @@ namespace Opc.Ua.Cloud.Library.Controllers
         [SwaggerResponse(statusCode: 200, type: typeof(string[]), description: "All OPC UA Information Model names and associated identifiers of the models found in the UA Cloud Library.")]
         public async Task<IActionResult> GetAllNamesandIdentifiersAsync()
         {
-            string[] results = await _database.GetAllNamesAndNodesets().ConfigureAwait(false);
+            string[] results = await _database.GetAllNamesAndNodesets(User.Identity.Name).ConfigureAwait(false);
             return new ObjectResult(results) { StatusCode = (int)HttpStatusCode.OK };
         }
 
@@ -103,7 +103,7 @@ namespace Opc.Ua.Cloud.Library.Controllers
                 return new ObjectResult("Could not parse identifier") { StatusCode = (int)HttpStatusCode.BadRequest };
             }
 
-            string[] types = await _database.GetAllTypes(identifier).ConfigureAwait(false);
+            string[] types = await _database.GetAllTypes(User.Identity.Name, identifier).ConfigureAwait(false);
             if ((types == null) || (types.Length == 0))
             {
                 return new ObjectResult("Failed to find nodeset types") { StatusCode = (int)HttpStatusCode.NotFound };
@@ -126,7 +126,7 @@ namespace Opc.Ua.Cloud.Library.Controllers
                 return new ObjectResult("Could not parse identifier") { StatusCode = (int)HttpStatusCode.BadRequest };
             }
 
-            string[] instances = await _database.GetAllInstances(identifier).ConfigureAwait(false);
+            string[] instances = await _database.GetAllInstances(User.Identity.Name, identifier).ConfigureAwait(false);
             if ((instances == null) || (instances.Length == 0))
             {
                 return new ObjectResult("Failed to find nodeset instances") { StatusCode = (int)HttpStatusCode.NotFound };
@@ -150,7 +150,7 @@ namespace Opc.Ua.Cloud.Library.Controllers
                 return new ObjectResult("Could not parse expanded node ID") { StatusCode = (int)HttpStatusCode.BadRequest };
             }
 
-            string nodeInfo = _database.GetNode(expandedNodeId, identifier);
+            string nodeInfo = _database.GetNode(expandedNodeId, identifier, User.Identity.Name);
             if (string.IsNullOrEmpty(nodeInfo))
             {
                 return new ObjectResult("Failed to find node information") { StatusCode = (int)HttpStatusCode.NotFound };
@@ -187,22 +187,24 @@ namespace Opc.Ua.Cloud.Library.Controllers
                 return new ObjectResult("Could not parse identifier") { StatusCode = (int)HttpStatusCode.BadRequest };
             }
 
+            UANameSpace uaNamespace = await _database.RetrieveAllMetadataAsync(nodeSetID, User.Identity.Name).ConfigureAwait(false);
+            if (uaNamespace == null)
+            {
+                return new ObjectResult("Failed to find nodeset metadata") { StatusCode = (int)HttpStatusCode.NotFound };
+            }
+
+            // if we got here, the user is allowed to download the nodeset
+
             if (nodesetXMLOnly)
             {
                 await _database.IncrementDownloadCountAsync(nodeSetID).ConfigureAwait(false);
                 return new ObjectResult(nodesetXml) { StatusCode = (int)HttpStatusCode.OK };
             }
 
-            UANameSpace uaNamespace = await _database.RetrieveAllMetadataAsync(nodeSetID).ConfigureAwait(false);
-            if (uaNamespace == null)
-            {
-                return new ObjectResult("Failed to find nodeset metadata") { StatusCode = (int)HttpStatusCode.NotFound };
-            }
-
-            uaNamespace.Nodeset.NodesetXml = nodesetXml.Blob;
-
             if (!metadataOnly)
             {
+                uaNamespace.Nodeset.NodesetXml = nodesetXml.Blob;
+
                 // Only count downloads with XML payload
                 await _database.IncrementDownloadCountAsync(nodeSetID).ConfigureAwait(false);
             }
@@ -232,7 +234,12 @@ namespace Opc.Ua.Cloud.Library.Controllers
                 return new ObjectResult("Failed to find nodeset") { StatusCode = (int)HttpStatusCode.NotFound };
             }
 
-            UANameSpace uaNamespace = await _database.RetrieveAllMetadataAsync(nodeSetID).ConfigureAwait(false);
+            UANameSpace uaNamespace = await _database.RetrieveAllMetadataAsync(nodeSetID, User.Identity.Name).ConfigureAwait(false);
+            if (uaNamespace == null)
+            {
+                return new ObjectResult("Failed to find nodeset metadata") { StatusCode = (int)HttpStatusCode.NotFound };
+            }
+
             uaNamespace.Nodeset.NodesetXml = nodesetXml.Blob;
 
             await _database.DeleteAllRecordsForNodesetAsync(nodeSetID).ConfigureAwait(false);
